+++ bootique-jersey-jakarta-client-junit5-wiremock/src/main/java/io/bootique/jersey/client/junit5/wiremock/WireMockTester.java
@@ -24,99 +24,60 @@ import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import io.bootique.BQCoreModule;
 import io.bootique.di.BQModule;
 import io.bootique.junit5.BQTestScope;
-import io.bootique.junit5.scope.BQAfterMethodCallback;
 import io.bootique.junit5.scope.BQAfterScopeCallback;
-import io.bootique.junit5.scope.BQBeforeMethodCallback;
 import io.bootique.junit5.scope.BQBeforeScopeCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.io.IOException;
-
 /**
- * A Bootique test tool that customises and manages Wiremock lifecycle: server start/stop, recording start/stop.
- * Should be annotated with {@link io.bootique.junit5.BQTestTool}. To enable the recording mode run tests with
- * {code}-Dbq.wiremock.recording{code} argument. Wiremock will proxy targetUrl and persist responses as mappings.
+ * A Bootique test tool that sets up and manages a WireMock "server". Each tester should be annotated with
+ * {@link io.bootique.junit5.BQTestTool}.
  *
  * @since 3.0
  */
-public class WireMockTester implements BQBeforeScopeCallback, BQAfterScopeCallback, BQBeforeMethodCallback, BQAfterMethodCallback {
+public abstract class WireMockTester<T extends WireMockTester<T>> implements BQBeforeScopeCallback, BQAfterScopeCallback {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(WireMockTester.class);
-    private static final File DEFAULT_FILES_ROOT = new File("src/test/resources/wiremock/");
+
+    protected WireMockConfiguration config;
+    protected volatile WireMockServer wiremockServer;
 
     /**
-     * If present, enables wiremock recording mode for all tests
-     */
-    public static final String RECORDING_PROPERTY = "bq.wiremock.recording";
-
-    private final WireMockUrlParts url;
-    private WireMockConfiguration config;
-    private File filesRoot;
-    private boolean recordingEnabled;
-
-    private volatile WireMockServer wiremockServer;
-
-    /**
+     * Creates a new mock tester with a real backend at the given URL.
+     *
      * @param targetUrl url of resource to mock, for example: "https://www.example.com/foo/bar"
      */
-    public static WireMockTester create(String targetUrl) {
-        return new WireMockTester(WireMockUrlParts.of(targetUrl));
+    public static WireMockRecordingTester recordingTester(String targetUrl) {
+        return new WireMockRecordingTester(WireMockUrlParts.of(targetUrl));
+    }
+
+    /**
+     * Creates a new mock tester whose behavior should be later configured with custom stubs.
+     */
+    public static WireMockStubbingTester stubbingTester() {
+        return new WireMockStubbingTester();
     }
 
     // keep for a while to allow early adopters to upgrade
     @Deprecated(since = "3.0.M2")
-    public static WireMockTester tester(String targetUrl) {
-        return create(targetUrl);
+    public static WireMockRecordingTester create(String targetUrl) {
+        return recordingTester(targetUrl);
     }
 
-    protected WireMockTester(WireMockUrlParts url) {
-        this.url = url;
-        this.recordingEnabled = System.getProperty(RECORDING_PROPERTY) != null;
-    }
-
-    /**
-     * Enables the recording mode for this tester. It will result in this tester going to the origin for the request
-     * data and storing it in a file. The same can be achieved for all testers at once by starting the tests JVM with
-     * {code}-Dbq.wiremock.recording{code} property.
-     */
-    public WireMockTester recordingEnabled() {
-        this.recordingEnabled = true;
-        return this;
+    // keep for a while to allow early adopters to upgrade
+    @Deprecated(since = "3.0.M2")
+    public static WireMockRecordingTester tester(String targetUrl) {
+        return recordingTester(targetUrl);
     }
 
     /**
-     * A builder method that sets an optional custom config for WireMockServer. Overrides all customization made to the
+     * A builder method that sets an optional custom config for WireMockServer. Overrides certain customization made to the
      * tester.
      */
-    public WireMockTester config(WireMockConfiguration config) {
+    public T config(WireMockConfiguration config) {
         this.config = config;
-
-        // reset all custom configs
-        this.filesRoot = null;
-        return this;
-    }
-
-    /**
-     * A builder method that establishes a local directory that will be used as a root for WireMock recording files.
-     * If not set, a default location will be picked automatically.
-     */
-    public WireMockTester filesRoot(File filesRoot) {
-        this.filesRoot = filesRoot;
-
-        // reset custom config override
-        this.config = null;
-        return this;
-    }
-
-    /**
-     * A builder method that establishes a local directory that will be used as a root for WireMock recording files.
-     * If not set, a default location will be picked automatically.
-     */
-    public WireMockTester filesRoot(String filesRoot) {
-        return filesRoot(new File(filesRoot));
+        return (T) this;
     }
 
     @Override
@@ -147,20 +108,11 @@ public class WireMockTester implements BQBeforeScopeCallback, BQAfterScopeCallba
     }
 
     /**
-     * Returns the target URL of the tester rewritten to access the WireMock proxy.
+     * Returns the URL of the internal WireMock server.
      */
-    public String getUrl() {
-        return ensureRunning().baseUrl() + url.getPath();
-    }
+    public abstract String getUrl();
 
-    /**
-     * The name of the tester-specific folder located in the wiremock scenarios base folder.
-     */
-    public String getFolder() {
-        return url.getTesterFolder();
-    }
-
-    private WireMockServer ensureRunning() {
+    protected WireMockServer ensureRunning() {
         if (wiremockServer == null) {
             synchronized (this) {
                 if (wiremockServer == null) {
@@ -173,49 +125,24 @@ public class WireMockTester implements BQBeforeScopeCallback, BQAfterScopeCallba
         return wiremockServer;
     }
 
-    private WireMockServer createServer() {
-        WireMockConfiguration config = createServerConfig();
-        return new WireMockServer(config);
+    protected WireMockConfiguration ensureServerConfig() {
+        return this.config != null ? this.config : createServerConfig();
     }
 
-    private WireMockConfiguration createServerConfig() {
-        if (this.config != null) {
-            return this.config;
-        }
-
-        // TODO: the default must be portable and should not assume Maven project structure
-        File filesRoot = this.filesRoot != null ? this.filesRoot : DEFAULT_FILES_ROOT;
-        File testerRoot = new File(filesRoot, url.getTesterFolder());
-
+    protected WireMockConfiguration createServerConfig() {
         return WireMockConfiguration
                 .wireMockConfig()
                 .dynamicPort()
-                .notifier(new Slf4jNotifier(true))
-                .usingFilesUnderDirectory(testerRoot.getAbsolutePath());
+                .notifier(new Slf4jNotifier(true));
     }
 
-    private void startServer() {
+    protected WireMockServer createServer() {
+        WireMockConfiguration config = ensureServerConfig();
+        return new WireMockServer(config);
+    }
+
+    protected void startServer() {
         wiremockServer.start();
-
-        if (recordingEnabled) {
-            LOGGER.info("Wiremock started in recording mode on port {}", wiremockServer.port());
-        } else {
-            LOGGER.info("Wiremock started on port {}", wiremockServer.port());
-        }
-    }
-
-    @Override
-    public void beforeMethod(BQTestScope scope, ExtensionContext context) {
-        if (recordingEnabled) {
-            // note that we are recording relative to the "baseUrl", not the full target URL
-            WireMockRecorder.startRecording(wiremockServer, url.getBaseUrl());
-        }
-    }
-
-    @Override
-    public void afterMethod(BQTestScope scope, ExtensionContext context) throws IOException {
-        if (recordingEnabled) {
-            WireMockRecorder.stopRecording(wiremockServer, context);
-        }
+        LOGGER.info("WireMock started on port {}", wiremockServer.port());
     }
 }