+++ bootique-jersey-jakarta-client-junit5-wiremock/src/test/java/io/bootique/jersey/client/junit5/wiremock/WireMockStubbingTesterIT.java
@@ -21,7 +21,6 @@ package io.bootique.jersey.client.junit5.wiremock;
 import io.bootique.BQRuntime;
 import io.bootique.Bootique;
 import io.bootique.jersey.client.HttpTargets;
-import io.bootique.jersey.client.junit5.wiremock.junit.WireMockRecordingTest;
 import io.bootique.jetty.junit5.JettyTester;
 import io.bootique.junit5.BQApp;
 import io.bootique.junit5.BQTest;
@@ -30,14 +29,16 @@ import jakarta.ws.rs.client.WebTarget;
 import jakarta.ws.rs.core.MediaType;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
 
 @BQTest
-public class WireMockTester_WithPathIT extends WireMockRecordingTest {
+public class WireMockStubbingTesterIT {
 
     @BQTestTool
-    static final WireMockRecordingTester tester = WireMockTester.recordingTester(SERVER_URL + "/p1");
-
+    static final WireMockStubbingTester tester = WireMockTester.stubbingTester()
+            .stub(get("/s1").willReturn(aResponse().withHeader("Content-Type", "text/plain").withBody("[s1]")))
+            .stub(get("/s2").willReturn(aResponse().withHeader("Content-Type", "text/plain").withBody("[s2]")));
 
     @BQApp(skipRun = true)
     static final BQRuntime app = Bootique.app()
@@ -46,18 +47,14 @@ public class WireMockTester_WithPathIT extends WireMockRecordingTest {
             .createRuntime();
 
     @Test
-    public void testTarget() {
+    public void test() {
         WebTarget target = app.getInstance(HttpTargets.class).newTarget("tester");
-        JettyTester.assertOk(target.request().get())
+        JettyTester.assertOk(target.path("s1").request().get())
                 .assertContentType(MediaType.TEXT_PLAIN)
-                .assertContent(c -> assertTrue(c.contains("get:p1")));
-    }
-
-    @Test
-    public void testSubTarget() {
-        WebTarget target = app.getInstance(HttpTargets.class).newTarget("tester").path("p11");
-        JettyTester.assertOk(target.request().get())
+                .assertContent("[s1]");
+        JettyTester.assertOk(target.path("s2").request().get())
                 .assertContentType(MediaType.TEXT_PLAIN)
-                .assertContent(c -> assertTrue(c.contains("get:p1:p11")));
+                .assertContent("[s2]");
+        JettyTester.assertNotFound(target.path("s3").request().get());
     }
 }