+++ bootique-jersey-jakarta-client-junit5-wiremock/src/test/java/io/bootique/jersey/client/junit5/wiremock/junit/WireMockRecordingTest.java
@@ -33,7 +33,7 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 
 @BQTest
-public abstract class BaseTest {
+public abstract class WireMockRecordingTest {
 
     protected static final String SERVER_URL = "http://localhost:16348";